state: Move Hooks into state package
