+++ internal/state/hooks.go
@@ -1,7 +1,6 @@
-package config
+package state
 
 import (
-	"github.com/canonical/microcluster/internal/state"
 	"github.com/canonical/microcluster/rest/types"
 )
 
@@ -9,34 +8,34 @@ import (
 // integrate with other tools.
 type Hooks struct {
 	// PreBootstrap is run before the daemon is initialized and bootstrapped.
-	PreBootstrap func(s *state.State, initConfig map[string]string) error
+	PreBootstrap func(s State, initConfig map[string]string) error
 
 	// PostBootstrap is run after the daemon is initialized and bootstrapped.
-	PostBootstrap func(s *state.State, initConfig map[string]string) error
+	PostBootstrap func(s State, initConfig map[string]string) error
 
 	// OnStart is run after the daemon is started.
-	OnStart func(s *state.State) error
+	OnStart func(s State) error
 
 	// PostJoin is run after the daemon is initialized, joined the cluster and existing members triggered
 	// their 'OnNewMember' hooks.
-	PostJoin func(s *state.State, initConfig map[string]string) error
+	PostJoin func(s State, initConfig map[string]string) error
 
 	// PreJoin is run after the daemon is initialized and joined the cluster but before existing members triggered
 	// their 'OnNewMember' hooks.
-	PreJoin func(s *state.State, initConfig map[string]string) error
+	PreJoin func(s State, initConfig map[string]string) error
 
 	// PreRemove is run on a cluster member just before it is removed from the cluster.
-	PreRemove func(s *state.State, force bool) error
+	PreRemove func(s State, force bool) error
 
 	// PostRemove is run on all other peers after one is removed from the cluster.
-	PostRemove func(s *state.State, force bool) error
+	PostRemove func(s State, force bool) error
 
 	// OnHeartbeat is run after a successful heartbeat round.
-	OnHeartbeat func(s *state.State) error
+	OnHeartbeat func(s State) error
 
 	// OnNewMember is run on each peer after a new cluster member has joined and executed their 'PreJoin' hook.
-	OnNewMember func(s *state.State) error
+	OnNewMember func(s State) error
 
-	// OnDaemonConfigUpdate
-	OnDaemonConfigUpdate func(s *state.State, config types.DaemonConfig) error
+	// OnDaemonConfigUpdate is a post-action hook that is run on all cluster members when any cluster member receives a local configuration update.
+	OnDaemonConfigUpdate func(s State, config types.DaemonConfig) error
 }
